Move drawing functions to 16bit colour - better theme colours
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@8134 8aca7d54-2c30-db11-9de9-000461428c89

+++ simfab.cc
@@ -2620,7 +2620,7 @@ void fabrik_t::new_month()
 
 
 // static !
-unsigned fabrik_t::status_to_color[5] = {COL_RED, COL_ORANGE, COL_GREEN, COL_YELLOW, COL_WHITE };
+uint8 fabrik_t::status_to_color[5] = {COL_RED, COL_ORANGE, COL_GREEN, COL_YELLOW, COL_WHITE };
 
 #define FL_WARE_NULL           1
 #define FL_WARE_ALLENULL       2
